Clarify default table name for model
It is less than obvious to the Dynogels newcomer that DynamoDB table names are inferred from Dynogels model names, but are lowercased and pluralized. This can waste a lot of time unless the reader happens upon a statement to this effect later in the readme. Consequently, I think it would be helpful to state this upfront at the reader's first exposure to the model and the define method.